requirements.txt
admin-totals==1.0.1
aggdraw==1.3.15
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.6.3
asgiref==3.5.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
bokeh==2.4.3
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.5
cycler==0.11.0
dataframe-image==0.1.1
decorator==5.1.1
defusedxml==0.7.1
df2img==0.2.5
diff-match-patch==20200713
dj-database-url==1.0.0
Django==4.0.2
django-admin-interface==0.18.6
django-admin-rangefilter==0.8.3
django-colorfield==0.6.3
django-cors-headers==4.0.0
django-filter==23.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==2.7.1
django-jazzmin==2.6.0
django-jet-reboot==1.3.1
django-queryable-properties==1.8.3
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
entrypoints==0.4
environ==1.0
environs==9.5.0
et-xmlfile==1.1.0
executing==0.8.3
Faker==15.1.1
fastjsonschema==2.15.3
fonttools==4.33.3
frozenlist==1.3.0
google-api-python-client==1.4.1
httplib2==0.20.4
idna==3.3
imageio==2.19.5
ipython==8.4.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.6.1
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
kaleido==0.2.1
kiwisolver==1.4.3
MarkupPy==1.14
MarkupSafe==2.1.1
marshmallow==3.15.0
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.2
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.5
numpy==1.22.2
oauth2client==4.1.3
odfpy==1.4.1
opencv-python==4.5.5.62
openpyxl==3.0.9
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.1
plotly==5.9.0
prompt-toolkit==3.0.30
psd-tools==1.9.21
psycopg2==2.9.5
pure-eval==0.2.2
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
PyJWT==2.7.0
pyparsing==3.0.7
pyperclip==1.8.2
pyrsistent==0.18.1
pytesseract==0.3.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
pywin32==304
PyYAML==6.0
pyzmq==23.2.0
requests==2.27.1
rsa==4.9
schedule==1.1.0
scikit-image==0.19.3
scipy==1.8.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.3.0
tablib==3.2.0
Telethon==1.24.0
tenacity==8.0.1
tifffile==2022.5.4
tinycss2==1.1.1
tornado==6.1
traitlets==5.3.0
typing_extensions==4.2.0
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.3
whitenoise==6.4.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.7.2
